Apply suggestions from latest review

+++ cuebot/src/main/java/com/imageworks/spcue/dispatcher/BookingQueue.java
@@ -87,7 +87,7 @@ public class BookingQueue implements QueueHealthCheck {
         healthyThreadPool.shutdown();
     }
 
-    public void setShutdownDrainMs(int shutdownDrainMs) {
+    public void setShutdownDrainMs(long shutdownDrainMs) {
         healthyThreadPool.setShutdownDrainMs(shutdownDrainMs);
     }
